samples/PyPDF2/requirements.txt
packaging==25.0
pillow==11.2.1
PyPDF2==3.0.1
pytesseract==0.3.13
